Add most of infrastructure for solver test

+++ pyviennacl/src/pyviennacl/linalg.py
@@ -352,7 +352,8 @@ def solve(A, B, tag):
         raise TypeError("B must be Matrix or Vector type")
 
     try:
-        return result_type(A.vcl_leaf.solve(B.vcl_leaf, tag.vcl_tag))
+        return result_type(A.vcl_leaf.solve(B.vcl_leaf, tag.vcl_tag),
+                           dtype = A.dtype)
     except AttributeError:
         raise TypeError("tag must be a supported solver tag!")
 Matrix.solve = solve # for convenience..
